% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_wins.R
\name{show_wins}
\alias{show_wins}
\title{This will get the win amounts and compare them to each other on a bar plot in a
single column.}
\usage{
show_wins(team, data)
}
\arguments{
\item{team}{This is the team you want information about and needs to be 'team'}

\item{data}{This is where you're pulling your info from (summary_tibble.csv)}
}
\value{
This returns a bar plot
}
\description{
This will get the win amounts and compare them to each other on a bar plot in a
single column.
}
