% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_summary.R
\name{win_summary}
\alias{win_summary}
\title{This function will summarize all the data gained from the website and get it ready
for any team that you select from the website.}
\usage{
win_summary(team, data)
}
\arguments{
\item{team}{This is the team that you want to get information about}

\item{data}{This is where you would draw information from}
}
\value{
This returns the data you will use to make the other graphs
}
\description{
This function will summarize all the data gained from the website and get it ready
for any team that you select from the website.
}
